Refactor AppShell into route-based application structure
