+++ apps/desktop/src/routes/__root.tsx
@@ -1,9 +1,51 @@
-import { createRootRoute } from "@tanstack/react-router";
+import { createRootRoute, Outlet } from "@tanstack/react-router";
 import { z } from "zod";
-import { AppShell } from "../AppShell";
 import type { SettingsSection } from "@/features/settings/SettingsSidebar";
 import type { InsightsSection } from "@/features/insights/InsightsPage";
 import type { AgentKind } from "../core/types";
+import { normalizePlatform } from "../core/platform";
+import { AppRuntimeBridge } from "../features/app/AppRuntimeBridge";
+import { GlobalShell } from "../features/app/GlobalShell";
+import { useAppNavigation } from "../features/app/useAppNavigation";
+
+const appPlatform = normalizePlatform(import.meta.env.TAURI_ENV_PLATFORM);
+
+function RootLayout() {
+  const {
+    route,
+    globalPage,
+    sidebarCollapsed,
+    setSidebarCollapsed,
+    isFullscreen,
+    message,
+    refreshJobs,
+    navigation,
+    navigateGlobal,
+    openSettings,
+  } = useAppNavigation();
+
+  return (
+    <>
+      <AppRuntimeBridge />
+      {route.kind === "workspace" || route.kind === "settings" ? (
+        <Outlet />
+      ) : (
+        <GlobalShell
+          active={globalPage}
+          entries={navigation}
+          collapsed={sidebarCollapsed}
+          fullscreen={isFullscreen}
+          platform={appPlatform}
+          message={message}
+          refreshJobs={refreshJobs}
+          onToggleSidebar={() => setSidebarCollapsed((value) => !value)}
+          onNavigate={navigateGlobal}
+          onSettings={openSettings}
+        />
+      )}
+    </>
+  );
+}
 
 const quotaWindowSchema = z.object({
   provider_id: z.string(),
@@ -54,6 +96,6 @@ const searchSchema = z.object({
 
 export const Route = createRootRoute({
   validateSearch: searchSchema,
-  component: AppShell,
+  component: RootLayout,
   notFoundComponent: () => <div>Not found</div>,
 });
